+++ src/engine/engineaux.h
